+++ packages/cli/src/databases/types.ts
@@ -1,5 +1,5 @@
 import type { INodeTypes } from 'n8n-workflow';
-import type { QueryRunner, ObjectLiteral } from 'typeorm';
+import { type QueryRunner, type ObjectLiteral, QueryFailedError } from 'typeorm';
 import type { Logger } from '@/Logger';
 import type { createSchemaBuilder } from './dsl';
 
@@ -59,3 +59,5 @@ export interface Migration extends Function {
 }
 
 export type InsertResult = Array<{ insertId: number }>;
+
+export { QueryFailedError } from 'typeorm/error/QueryFailedError';
